agent.py
from langchain.chat_models import ChatOpenAI
from langchain.agents import initialize_agent, AgentType
from tools import sentiment_tool, coherence_tool
from dotenv import load_dotenv
import os

# Carrega as variáveis de ambiente do arquivo .env
load_dotenv()

def create_agent():
    llm = ChatOpenAI(
        temperature=0,
        openai_api_key=os.getenv("OPENAI_API_KEY"),
        model_name="gpt-3.5-turbo",
    )

    tools = [sentiment_tool, coherence_tool]

    agent = initialize_agent(
        tools=tools,
        llm=llm,
        agent=AgentType.ZERO_SHOT_REACT_DESCRIPTION,
        verbose=True
    )

    return agent
