requirements.txt
openai>=0.27.0
langchain>=0.0.300
streamlit>=1.10.0
python-dotenv
openai==0.27.0


